refactor(sql): remove gnrsqlmodel_ prefix from sub-package module names
Rename all modules in the gnrsqlmodel/ sub-package to drop the redundant
gnrsqlmodel_ prefix (e.g. gnrsqlmodel_helpers.py → helpers.py).
Update all intra-package imports, docstrings and README accordingly.

Ref #498
