+++ gnrpy/gnr/sql/gnrsqlmodel/helpers.py
