+++ gnrpy/gnr/sql/gnrsqlmodel/obj.py
